Refactor tests into unit vs integration test
- Also split up the tests for BraveAlerterConfigurator to have a
  separate file per method being tested. This will make for smaller,
  easier to manage files that are easier to read and easier to find
  the particular tests that you're looking for
- Follows the same pattern as I'm using in Buttons and Brave Alert
  App
