+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "lint": "eslint .",
     "test": "NODE_ENV=test mocha --timeout 10000 \"test/**/*.js\"",
-    "integrationtest": "NODE_ENV=test mocha --timeout 10000 \"test/serverTest.js\"",
-    "unittest": "NODE_ENV=test mocha --timeout 10000 \"test/testXeThruStateMachine.js\"",
+    "integrationtest": "NODE_ENV=test mocha --timeout 10000 \"test/integration/**/*.js\"",
+    "unittest": "NODE_ENV=test mocha --timeout 10000 \"test/unit/**/*.js\"",
     "smoketest": "node -e 'require(\"./smokeTest.js\")'"
   },
   "eslintIgnore": [
