+++ test/unit/testXeThruStateMachine.js
@@ -3,13 +3,13 @@ const { describe, it } = require('mocha')
 const moment = require('moment')
 
 const expect = chai.expect
-const STATE = require('../SessionStateEnum.js')
-const XETHRU_STATE = require('../SessionStateXethruEnum.js')
-const OD_FLAG_STATE = require('../SessionStateODFlagEnum.js')
-const DOOR_STATE = require('../SessionStateDoorEnum.js')
-const StateMachine = require('../XeThruStateMachine.js')
-const Session = require('../Session.js')
-const Location = require('../Location.js')
+const STATE = require('../../SessionStateEnum.js')
+const XETHRU_STATE = require('../../SessionStateXethruEnum.js')
+const OD_FLAG_STATE = require('../../SessionStateODFlagEnum.js')
+const DOOR_STATE = require('../../SessionStateDoorEnum.js')
+const StateMachine = require('../../XeThruStateMachine.js')
+const Session = require('../../Session.js')
+const Location = require('../../Location.js')
 
 function setupDB(location_data = new Location(), session = new Session(), is_overdose_suspected = false) {
   return {
